small fixes

+++ api/controllers/weekController.ts
@@ -9,6 +9,19 @@ interface Week {
   filledSlots: number;
 }
 
+const checkWeekExists = async(weekDate: string) =>
+{
+	const weekExists = await prisma.week.findMany({
+		where: { weekDate: weekDate },
+		include: { jobs: true }
+	});
+
+	if(weekExists)
+		return true;
+
+	return false;
+};
+
 weeksRouter
 .get("/", async(req, res) => {
 	try {
@@ -23,27 +36,34 @@ weeksRouter
 		return res.status(200).json(allWeeks);
 	} catch (error) {
 		console.log("error ===", error);
-		res.status(404).json({ error: true, msg: error });
+		return res.status(404).json({ error: true, msg: error });
 	}
 })
 .post("/", async(req, res) => {
 	try {
 		const { weekDate, openSlots, filledSlots = 0 } = req.body;
 
+		const weekExists = await checkWeekExists(weekDate);
+
+		if(weekExists)
+		{
+			return res.status(400).json("This week date already exists");
+			return;
+		}
+
 		const week: Week = await prisma.week.create({
 			data: {
 				weekDate: weekDate,
 				openSlots: openSlots,
-				filledSlots: filledSlots,
-			},
+				filledSlots: filledSlots
+			}
 		});
 
-		return res
-		.status(200)
-		.json({ msg: "Week added successfully!", data: week });
+		return res.status(200).json({ msg: "Week added successfully!", data: week });
+
 	} catch (error) {
 		console.log("error ===", error);
-		res.status(404).json({ error: true, msg: error });
+		return res.status(404).json({ error: true, msg: error });
 	}
 })
 .get("/:id", async(req, res) => {
@@ -52,31 +72,39 @@ weeksRouter
 
 		const week = await prisma.week.findUnique({
 			where: { id: parseInt(id) },
-			include: { jobs: true },
+			include: { jobs: true }
 		});
 
 		if(!week) {
-			res.status(404).json({ error: true, msg: "Cannot find this week :(" });
+			return res.status(404).json({ error: true, msg: "Cannot find this week :(" });
 			return;
 		}
 
-		res.status(200).json({ msg: "week added successfully!", data: week });
+		return res.status(200).json(week);
 	} catch (error) {
 		console.log("error ===", error);
-		res.status(404).json({ error: true, msg: error });
+		return res.status(404).json({ error: true, msg: error });
 	}
 })
 .delete("/:id", async(req, res) => {
 	try {
 		const { id } = req.params;
 
+		const weekExists = await prisma.week.findUnique({ where: { id: parseInt(id) } });
+
+		if(!weekExists)
+		{
+			res.status(400).json("This week date doesn't exists");
+			return;
+		}
+
 		await prisma.week.delete({ where: { id: parseInt(id) } });
 
-		return res
-		.status(200)
-		.json({ msg: "Week has been deleted successfully" });
-	} catch (error) {
-		res.status(404).json({ msg: "we have an error", error: error });
+		return res.status(200).json({ msg: "Week has been deleted successfully" });
+
+	} catch (error)
+	{
+		return res.status(404).json({ msg: "we have an error", error: error });
 	}
 })
 .patch("/:id", async(req, res) => {
@@ -84,18 +112,25 @@ weeksRouter
 		const { id } = req.params;
 		const { weekDate, openSlots, filledSlots = 0 } = req.body;
 
+		const weekExists = await checkWeekExists(weekDate);
+
+		if(weekExists)
+		{
+			res.status(400).json("This week already exists! Please check the params and or week date");
+			return;
+		}
+
 		const week = await prisma.week.update({
 			where: { id: parseInt(id) },
 			data: {
 				weekDate: weekDate,
 				openSlots: openSlots,
-				filledSlots: filledSlots,
-			},
+				filledSlots: filledSlots
+			}
 		});
 
-		return res
-		.status(200)
-		.json({ msg: "Week updated successfully!", data: week });
+		return res.status(200).json({ msg: "Week updated successfully!", data: week });
+
 	} catch (error) {
 		console.log("error ===", error);
 		res.status(404).json({ error: true, msg: error });
